feat(os): add shared uefi kernel base
