+++ code/packages/rust/Cargo.toml
@@ -19,7 +19,6 @@ members = [
     "suffix-tree",
     "rope",
     "blas-library",
-    "boot-protocol",
     "bootloader",
     "brainfuck",
     "branch-predictor",
@@ -96,15 +95,13 @@ members = [
     "ml-framework-tf",
     "ml-framework-torch",
     "os-kernel",
+    "os-kernel-simulator",
     "parallel-execution-engine",
     "parser",
     "perceptron",
     "pipeline",
     "progress-bar",
     "riscv-simulator",
-    "riscv-rt",
-    "riscv-qemu-bootloader",
-    "riscv-qemu-kernel",
     "rom-bios",
     "ruby-lexer",
     "ruby-parser",
@@ -154,7 +151,6 @@ members = [
     "md5",
     "sha1",
     "uuid",
-    "uart-16550",
     "wasm-leb128",
     "wasm-module-parser",
     "wasm-opcodes",
