--- code/packages/rust/boot-protocol/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "boot-protocol"
-version = "0.1.0"
-edition = "2021"
-description = "Shared boot contract for the real QEMU boot path"
-
