--- code/packages/rust/boot-protocol/required_capabilities.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/adhithyan15/coding-adventures/main/code/specs/schemas/required_capabilities.schema.json",
-  "version": 1,
-  "package": "rust/boot-protocol",
-  "capabilities": [],
-  "justification": "Pure in-memory data definitions for the bootloader and kernel handoff contract."
-}
