+++ code/packages/rust/os-kernel-simulator/BUILD
@@ -1,0 +1,1 @@
+cargo test -p os-kernel -- --nocapture
