+++ code/packages/rust/os-kernel-simulator/BUILD_windows
@@ -1,0 +1,1 @@
+cargo test -p os-kernel -- --nocapture
