+++ code/packages/rust/os-kernel-simulator/required_capabilities.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "https://raw.githubusercontent.com/adhithyan15/coding-adventures/main/code/specs/schemas/required_capabilities.schema.json",
+  "version": 1,
+  "package": "rust/os-kernel-simulator",
+  "capabilities": [],
+  "justification": "Simulated OS kernel. Implements process scheduling, system calls, and memory management in-memory without real OS interaction."
+}
