+++ code/packages/rust/os-kernel/BUILD
@@ -1,1 +1,1 @@
-cargo test -p os-kernel -- --nocapture
+rust_package()
