+++ code/packages/rust/os-kernel/required_capabilities.json
@@ -3,5 +3,5 @@
   "version": 1,
   "package": "rust/os-kernel",
   "capabilities": [],
-  "justification": "Simulated OS kernel. Implements process scheduling, system calls, and memory management in-memory without real OS interaction."
+  "justification": "Freestanding UEFI-loaded OS kernel. No host-side capabilities are used by the kernel at runtime."
 }
