--- code/packages/rust/riscv-qemu-bootloader/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "riscv-qemu-bootloader"
-version = "0.1.0"
-edition = "2021"
-description = "Custom Rust bootloader scaffold for QEMU RISC-V bring-up"
-build = "build.rs"
-
-[dependencies]
-boot-protocol = { path = "../boot-protocol" }
-riscv-rt = { path = "../riscv-rt" }
-uart-16550 = { path = "../uart-16550" }
