--- code/packages/rust/riscv-qemu-bootloader/required_capabilities.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/adhithyan15/coding-adventures/main/code/specs/schemas/required_capabilities.schema.json",
-  "version": 1,
-  "package": "rust/riscv-qemu-bootloader",
-  "capabilities": [],
-  "justification": "Freestanding bootloader scaffold for QEMU RISC-V with only memory-mapped device access."
-}
