--- code/packages/rust/riscv-qemu-bootloader/src/entry.S
@@ -1,9 +1,0 @@
-.section .text._start,"ax"
-.globl _start
-
-_start:
-    la sp, __stack_top
-    call bootloader_main
-
-1:
-    j 1b
