--- code/packages/rust/riscv-qemu-kernel/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "riscv-qemu-kernel"
-version = "0.1.0"
-edition = "2021"
-description = "First real no_std kernel for QEMU RISC-V bring-up"
-build = "build.rs"
-
-[dependencies]
-boot-protocol = { path = "../boot-protocol" }
-riscv-rt = { path = "../riscv-rt" }
-uart-16550 = { path = "../uart-16550" }
