--- code/packages/rust/riscv-qemu-kernel/required_capabilities.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/adhithyan15/coding-adventures/main/code/specs/schemas/required_capabilities.schema.json",
-  "version": 1,
-  "package": "rust/riscv-qemu-kernel",
-  "capabilities": [],
-  "justification": "Freestanding bare-metal kernel for QEMU RISC-V with no host OS interaction."
-}
