--- code/packages/rust/riscv-qemu-kernel/src/entry.S
@@ -1,9 +1,0 @@
-.section .text._start,"ax"
-.globl _start
-
-_start:
-    la sp, __stack_top
-    call kernel_main
-
-1:
-    j 1b
