--- code/packages/rust/riscv-rt/required_capabilities.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/adhithyan15/coding-adventures/main/code/specs/schemas/required_capabilities.schema.json",
-  "version": 1,
-  "package": "rust/riscv-rt",
-  "capabilities": [],
-  "justification": "Bare-metal runtime helpers for BSS initialization, instruction fences, and idle spinning."
-}
