--- code/packages/rust/uart-16550/required_capabilities.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/adhithyan15/coding-adventures/main/code/specs/schemas/required_capabilities.schema.json",
-  "version": 1,
-  "package": "rust/uart-16550",
-  "capabilities": [],
-  "justification": "Memory-mapped UART register access helpers for bare-metal bring-up."
-}
